rewrite to sequential flarefind
