+++ laff/__init__.py
@@ -1,2 +1,2 @@
-from .laff import findFlares, sequential_findFlares, fitContinuum, fitFlares, fitGRB, plotGRB
+from .laff import findFlares, fitContinuum, fitFlares, fitGRB, plotGRB
 from .lightcurve_import import lcimport
