+++ laff/flarefinding/__init__.py
@@ -1,12 +1,1 @@
-from .flarefinding import (
-    possible_flares,
-    _remove_Duplicates,
-    _check_AverageNoise,
-    _check_FluxIncrease,
-    _check_PulseShape,
-    _check_AboveContinuum
-    )
-
-from .sequential_flarefinding import (
-    findFlares
-)
+from .flarefinding import findFlares
